requirements.txt
python-binance==0.5.4
gdax==1.0.6